Rename the deployment component to not conflict with the composition
