+++ src/components/index.ts
@@ -239,7 +239,7 @@ export { default as TaskRunsSort } from './TaskRunsSort.vue'
 export { default as TimeSpanFilter } from './TimeSpanFilter.vue'
 export { default as TimezoneSelect } from './TimezoneSelect.vue'
 export { default as ToastFlowRunCreate } from './ToastFlowRunCreate.vue'
-export { default as UseDeployment } from './UseDeployment.vue'
+export { default as UseDeploymentSlot } from './UseDeploymentSlot.vue'
 export { default as VariableCreateModal } from './VariableCreateModal.vue'
 export { default as VariableEditModal } from './VariableEditModal.vue'
 export { default as VariableMenu } from './VariableMenu.vue'
